u32 -> u64 addresses
Since ranges are being used with an open end, we need to support
one more than u32 MAX. Hence, 64 bits are needed. DataChunks are
using the same amount of memory anyway.

+++ src/srecord/data_chunk.rs
@@ -2,26 +2,25 @@ use std::ops::Range;
 
 use crate::srecord::get::Get;
 
-// TODO: pub?
 #[derive(Debug, PartialEq, Eq)]
 pub struct DataChunk {
     /// Inclusive address of the start of the data chunk.
-    pub address: u32,
+    pub address: u64,
     /// Raw data of data chunk.
     pub data: Vec<u8>,
 }
 
-impl Get<u32> for DataChunk {
+impl Get<u64> for DataChunk {
     type Output = u8;
 
-    fn get(&self, address: u32) -> Option<&Self::Output> {
+    fn get(&self, address: u64) -> Option<&Self::Output> {
         match address.checked_sub(self.address) {
             Some(index) => self.data.get(index as usize),
             None => None,
         }
     }
 
-    fn get_mut(&mut self, address: u32) -> Option<&mut Self::Output> {
+    fn get_mut(&mut self, address: u64) -> Option<&mut Self::Output> {
         match address.checked_sub(self.address) {
             Some(index) => self.data.get_mut(index as usize),
             None => None,
@@ -29,10 +28,10 @@ impl Get<u32> for DataChunk {
     }
 }
 
-impl Get<Range<u32>> for DataChunk {
+impl Get<Range<u64>> for DataChunk {
     type Output = [u8];
 
-    fn get(&self, address_range: Range<u32>) -> Option<&Self::Output> {
+    fn get(&self, address_range: Range<u64>) -> Option<&Self::Output> {
         match address_range.start.checked_sub(self.address) {
             Some(start_index) => match address_range.end.checked_sub(self.address) {
                 Some(end_index) => self.data.get(start_index as usize..end_index as usize),
@@ -42,7 +41,7 @@ impl Get<Range<u32>> for DataChunk {
         }
     }
 
-    fn get_mut(&mut self, address_range: Range<u32>) -> Option<&mut Self::Output> {
+    fn get_mut(&mut self, address_range: Range<u64>) -> Option<&mut Self::Output> {
         match address_range.start.checked_sub(self.address) {
             Some(start_index) => match address_range.end.checked_sub(self.address) {
                 Some(end_index) => self.data.get_mut(start_index as usize..end_index as usize),
